+++ config/versiones.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    'IEE'           => '1.14.0',
-    'BhagamApps'    => '1.14.0',
+    'IEE'           => '1.14.1',
+    'BhagamApps'    => '1.14.1',
     'User'          => '2.3.0',
     'Inventario'    => '2.10.5',
-    'Apps'          => '1.5.0',
+    'Apps'          => '1.5.1',
     'CrudGenerator' => '1.1.0',
 ];